change type to float

+++ simple_nn_runner.py
@@ -73,13 +73,13 @@ X_train, X_test, y_train, y_test = train_test_split(data[train_col_mask],
                                                    )
 
 
-train_features = torch.tensor(X_train.values)
-test_features = torch.tensor(X_test.values)
+train_features = torch.tensor(X_train.values.astype(np.float32))
+test_features = torch.tensor(X_test.values.astype(np.float32))
 
 print(train_features.shape)
 print(le.transform(y_train).shape)
-train_loader = TensorDataset(train_features, torch.tensor(le.transform(y_train)))
-test_loader = TensorDataset(test_features, torch.tensor(le.transform(y_test)))
+train_loader = TensorDataset(train_features, torch.tensor(le.transform(y_train).astype(np.float32)))
+test_loader = TensorDataset(test_features, torch.tensor(le.transform(y_test).astype(np.float32)))
 
 ##########################
 ### Model Train Funcs. ###
